restructure the validators module
