+++ hydrolib/core/dflowfm/validators/__init__.py
@@ -1,0 +1,5 @@
+"""Shared validator mixins for D-Flow FM block models."""
+
+from hydrolib.core.dflowfm.validators.fields import CoordinateValidator
+
+__all__ = ["CoordinateValidator"]
